# PS-drone.project-main/PS-drone.project-main/project/models/parts.js
const mongoose = require('mongoose');

const PartSchema = new mongoose.Schema({
    name: {
        type: String,
        required: true
    },
    type: {
        type: String,
        required: true
    },
    quantity: {
        type: Number,
        required: true
    },
    createdAt: {
        type: Date,
        default: Date.now
    }
});

module.exports = mongoose.model('part', PartSchema);